server

+++ controllers/seniorIdApplicationControllers/deleteSeniorIdApplication.js
@@ -1,36 +1,35 @@
 import pool from '../../config/db.js';
-import { deleteMultipleFromCloudinary } from '../../utils/cloudinaryUtils.js';
+import { cleanupLocalStorageUploads } from '../../utils/fileUtils.js';
 
 
 export const deleteSeniorIdApplication = async (req, res) => {
-  try {
-    const { seniorCitizenId } = req.params;
 
-    // DELETE IMAGES FROM CLOUDINARY
-    const [applicantRows] = await pool.query(`
-      SELECT 
-        senior_citizen_photo_id_public_Id as seniorCitizenPhotoPublicId,
-        senior_citizen_signature_public_id as seniorCitizenSignaturePublicId
-      FROM senior_citizen_id_applications 
+  const connection = await pool.getConnection();
+
+  try {
+    await connection.beginTransaction();
+
+    //////////////////////////////////////////////////////////////////////
+
+    const seniorCitizenId = req.params.seniorCitizenId;
+
+    //////////////////////////////////////////////////////////////////////
+
+    // GET UPLOADED FILES
+
+    const [uploadedFilesRows] = await connection.query(`
+      SELECT
+        senior_citizen_photo_id_url as seniorCitizenPhotoId,
+        senior_citizen_signature_url as seniorCitizenSignature
+      FROM senior_citizen_id_applications
       WHERE senior_citizen_id = ?`,
       [seniorCitizenId]
     );
 
-    const publicIdsToDelete = [];
-
-    if (applicantRows?.length > 0) {applicantRows
-      publicIdsToDelete.push(
-        ...Object.values(applicantRows[0]).filter(Boolean)
-      );
-    }
-
-    if (publicIdsToDelete.length > 0) {
-      await deleteMultipleFromCloudinary(publicIdsToDelete);
-      console.log(`Cleaned up ${publicIdsToDelete.length} images from Cloudinary`);
-    }
+    //////////////////////////////////////////////////////////////////////
 
     // GET RESIDENT ID
-    const [residentIdRows] = await pool.query(`
+    const [residentRows] = await pool.query(`
       SELECT 
         resident_id as residentId
       FROM senior_citizen_id_applications 
@@ -38,18 +37,65 @@ export const deleteSeniorIdApplication = async (req, res) => {
       [seniorCitizenId]
     );
 
-    const residentId = residentIdRows[0].residentId;
+    const residentId = residentRows[0].residentId;
 
-    // DELETE APPLICANT INFORMATION
-    await pool.query(
-      `DELETE FROM population WHERE resident_id = ?`, [residentId]
-    );
-    await pool.query(
-      `DELETE FROM professional_information WHERE resident_id = ?`, [residentId]
-    );
-    await pool.query(
-      `DELETE FROM contact_information WHERE resident_id = ?`, [residentId]
-    );
+    //////////////////////////////////////////////////////////////////////
+
+    // DELETE RELATED RECORDS ONLY IF TEMPORARY RESIDENT
+
+    if (residentId?.startsWith('T-RID')) {
+      await pool.query(
+        `DELETE FROM population WHERE resident_id = ?`, [residentId]
+      );
+      await pool.query(
+        `DELETE FROM professional_information WHERE resident_id = ?`, [residentId]
+      );
+      await pool.query(
+        `DELETE FROM contact_information WHERE resident_id = ?`, [residentId]
+      );
+    } else {
+      const [applications] = await connection.query(`
+        SELECT 'pwd' AS type FROM pwd_id_applications WHERE resident_id = ?
+        UNION
+        SELECT 'solo_parent' FROM solo_parent_id_applications WHERE resident_id = ?`,
+        [residentId, residentId]
+      );
+
+      if (applications.length === 0) {
+        // PROFESSIONAL INFORMATION
+        await connection.query(`
+          UPDATE professional_information
+          SET 
+            employment_status = NULL,
+            employment_category = NULL
+          WHERE resident_id = ?
+        `, [residentId]);
+
+        // CONTACT INFORMATION
+        await connection.query(`
+          UPDATE contact_information
+          SET 
+            street = NULL,
+            barangay = NULL,
+            telephone_number = NULL,
+            email_address = NULL
+          WHERE resident_id = ?
+        `, [residentId]);
+
+        // GOVERNMENT IDs
+        await connection.query(`
+          UPDATE government_ids
+          SET 
+            sss = NULL,
+            gsis = NULL,
+            pagibig = NULL,
+            philsys = NULL
+          WHERE resident_id = ?
+        `, [residentId]);
+      }  
+    }
+
+    //////////////////////////////////////////////////////////////////////
 
     // DELETE SOLO PARENT ID APPLICATION
     await pool.query(
@@ -57,16 +103,31 @@ export const deleteSeniorIdApplication = async (req, res) => {
       [seniorCitizenId]
     );
 
+    //////////////////////////////////////////////////////////////////////
+
+    // DELETE IMAGE FROM LOCAL STORAGE
+
+    await cleanupLocalStorageUploads({
+      photoId: uploadedFilesRows[0].seniorCitizenPhotoId,
+      signature: uploadedFilesRows[0].seniorCitizenSignature
+    });
+
+    //////////////////////////////////////////////////////////////////////
+
     return res.status(200).json({ 
       success: true,
       message: 'Application Deleted Successfully!',
-      seniorCitizenId: seniorCitizenId
+      seniorCitizenId
     });
   } catch (error) {
+    await connection.rollback();
     console.error('Error deleting application:', error);
-    res.status(500).json({ 
+
+    res.status(500).json({
       success: false,
       message: 'Server error'
     });
+  } finally {
+    connection.release();
   }
 };
